Move `Kinetics` into `kinetics/__init__.py`
